// settings/karabiner/types.ts
export interface KarabinerRules {
  description?: string;
  manipulators?: Manipulator[];
}

export interface SimpleKarabinerRules {
  from?: From;
  to?: To[];
}

export interface Manipulator {
  description?: string;
  type: "basic";
  from: From;
  to?: To[];
  to_after_key_up?: To[];
  to_if_alone?: To[];
  parameters?: Parameters;
  conditions?: Conditions[];
  to_if_held_down?: To[];
  to_delayed_action?: DelayedAction;
}

export interface DelayedAction {
  to_if_invoked?: DelayedActionSetVar[];
  to_if_canceled?: DelayedActionSetVar[];
}

export interface DelayedActionSetVar {
  set_variable?: {
    name: string;
    value: boolean | number | string;
  };
}

export interface Parameters {
  "basic.simultaneous_threshold_milliseconds"?: number;
  "basic.to_if_alone_timeout_milliseconds"?: number;
  "basic.to_if_held_down_threshold_milliseconds"?: number;
}

type Conditions =
  | FrontMostApplicationCondition
  | DeviceCondition
  | KeybaordTypeCondition
  | InputSourceCondition
  | VaribaleCondition
  | EventChangedCondition;
// ...
export interface From {
  key_code?: KeyCode;
  simultaneous?: SimultaneousFrom[];
  simultaneous_options?: SimultaneousOptions;
  modifiers?: Modifiers;
  apple_vendor_top_case_key_code?: KeyCode;
}

export interface Modifiers {
  optional?: string[];
  mandatory?: string[];
}

export interface To {
  key_code?: KeyCode;
  modifiers?: string[];
  shell_command?: string;
  set_variable?: {
    name: string;
    value: boolean | number | string;
  };
  mouse_key?: MouseKey;
  pointing_button?: string;
  /**
   * Power Management plugin
   * @example: sleep system
   * @see: {@link https://karabiner-elements.pqrs.org/docs/json/complex-modifications-manipulator-definition/to/software_function/iokit_power_management_sleep_system/}
   */
  software_function?: SoftwareFunction;
  apple_vendor_top_case_key_code?: KeyCode;
  lazy?: boolean;
  select_input_source?: InputSource,
}